pyproject.toml
[project]
name = "gpt-terminal-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.0.13",
    "langchain-chroma>=0.2.4",
    "langchain-community>=0.3.26",
    "langchain-core>=0.3.67",
    "langchain-openai>=0.3.27",
    "openai>=1.91.0",
    "python-dotenv>=1.1.1",
    "tiktoken>=0.9.0",
    "uuid>=1.30",
    "vlite>=0.2.7",
]
